+++ src/js/constants/constants.js
@@ -1,0 +1,5 @@
+const Constants = {
+  "version": "0.4.0"
+}
+
+export default Constants;