+++ src/js/constants/constants.module.js
@@ -1,0 +1,9 @@
+import Constants from "./constants";
+
+(() => {
+  "use strict";
+
+  angular.module("modwatchuploader.constants", [])
+    .constant("Constants", Constants);
+
+})();