+++ src/js/files/filesService.js
@@ -5,7 +5,7 @@ export default FilesService;
 function FilesService($q) {
   return {
     getFiles: {
-      mo: (filename) => {
+      mo(filename) {
         let deferred = $q.defer();
         ipc.send(filename ? "mo.getFilesNoDialog" : "mo.getFiles", filename);
         ipc.on("filesread", (files) => {
@@ -14,7 +14,7 @@ function FilesService($q) {
         return deferred.promise;
       },
       nmm: {
-        plugins: (filename) => {
+        plugins(filename) {
           let deferred = $q.defer();
           ipc.send(noDialog ? "nmm.getPluginsFileNoDialog" : "nmm.getPluginsFile", filename);
           ipc.on("filesread", (files) => {
@@ -22,7 +22,7 @@ function FilesService($q) {
           });
           return deferred.promise;
         },
-        ini: (filename) => {
+        ini(filename) {
           let deferred = $q.defer();
           ipc.send(filename ? "nmm.getIniFilesNoDialog" : "nmm.getIniFiles", filename);
           ipc.on("filesread", (files) => {