+++ src/js/login/login.module.js
@@ -1,0 +1,9 @@
+import LoginCtrl from "./loginCtrl";
+
+(() => {
+  "use strict";
+
+  angular.module("modwatchuploader.login", [])
+    .controller("LoginCtrl", LoginCtrl);
+
+})();