+++ src/js/persistence/persistence.module.js
@@ -1,0 +1,9 @@
+import PersistenceService from "./persistenceService";
+
+(() => {
+  "use strict";
+
+  angular.module("modwatchuploader.persistence", [])
+    .factory("PersistenceService", PersistenceService);
+
+})();