+++ src/js/routes/routeConfig.js
@@ -5,6 +5,18 @@ export default RouteConfig;
 function RouteConfig($routeProvider) {
   $routeProvider
     .when("/", {
+      templateUrl: "login.template.html",
+      controller: "LoginCtrl",
+      controllerAs: "vm",
+      bindToController: true
+    })
+    .when("/login", {
+      templateUrl: "login.template.html",
+      controller: "LoginCtrl",
+      controllerAs: "vm",
+      bindToController: true
+    })
+    .when("/upload", {
       templateUrl: "upload.template.html",
       controller: "UploadCtrl",
       controllerAs: "vm",